Extract MatchPlayer into own type
